backend/pyproject.toml
[tool.poetry]
name = "nezbank"
version = "0.1.0"
description = "some description"
authors = ["selivanov"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
django-extensions = "^3.1.5"
sentry-sdk = "^1.5.1"
dj-rest-auth = "^2.1.11"
drf-yasg = "^1.20.0"
djangorestframework-simplejwt = "^4.8.0"
psycopg2-binary = "2.8.5"
factory-boy = "^3.2.1"
django-cors-headers = "^3.10.1"
requests = "^2.27.1"
xmltodict = "^0.12.0"
celery = "^5.2.3"
django-celery-beat = "^2.2.1"
redis = "^4.1.0"
django-role-permissions = "^3.1.1"
flower = "^1.0.0"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
